refactor(schedule-details): rename the metrics chart to the runs chart
The component plots a schedule's runs on a time axis, so "metrics" named
neither what it draws nor what an operator looks for. Rename the directory,
files, symbol, and the chart region's accessible name to say runs instead.
